Change remove trigger to cancel trigger.
In order to simplify and clearly communicate to the user of the
middleware, remove is changed to cancel, and instead of creating a
trigger and then creating an action for the trigger, it is consolidated
into one step.

This means for the user of the library to cancel a trigger, they will
have to access the action.trigger object to pass in to the cancelTrigger
function.